requirements.txt
annotated-types==0.7.0
argcomplete==3.6.2
asgiref==3.9.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
black==25.1.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.3
click==8.1.7
cloud-init==25.1.4
cloudpathlib==0.21.1
colorama==0.4.4
command-not-found==0.3
confection==0.1.5
configobj==5.0.6
constantly==15.1.0
coverage==7.10.5
cryptography==3.4.8
cymem==2.0.11
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.7
django-admin-interface==0.19.1
django-colorfield==0.9.0
django-cors-headers==4.3.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
djangorestframework==3.14.0
drf-nested-routers==0.95.0
exceptiongroup==1.3.0
flake8==7.3.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.1.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.5.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
langcodes==3.5.0
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==8.10.0
murmurhash==1.0.13
mypy==1.17.1
mypy_extensions==1.1.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
pipx==1.7.1
platformdirs==4.4.0
pluggy==1.6.0
preshed==3.0.10
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.14.0
pycurl==7.44.1
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-flask==1.3.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2022.1
PyYAML==5.4.1
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.1
scipy==1.15.3
SecretStorage==3.3.1
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
smart_open==7.3.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
sqlparse==0.5.3
srsly==2.5.1
SuperClaude==4.0.8
systemd-python==234
thinc==8.3.6
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
Twisted==22.1.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
userpath==1.9.2
uv==0.8.14
wadllib==1.3.6
wasabi==1.1.3
weasel==0.4.1
Werkzeug==2.3.7
wrapt==1.17.3
zipp==1.0.0
zope.interface==5.4.0
